rp2pio: Add offset & pc (program counter) properties
If StateMachine.run needs to contain a JMP instruction, the offset property
can be used to correctly encode the target address.

StateMachine.pc is useful for debugging, as it can give an overall
statistical idea of which state machine instructions are executing, or
show the instruction where the machine is stalled.