Work around an sdk bug
This was already reported & fixed upstream, but the fix will not be
in a released SDK until 2.0.1. It's easy enough to use our own workaround.

Testing: without this bug fix, rotaryencoder did not work on a pico 2

Note: the cut&paste code was reindented by uncrustify, giving it a slightly
different style than the upstream sdk code.