MDNS hostname match DHCP. Fix collision mangling
MDNS defaults to the hostname used by DHCP/LWIP since it is now
unique. This makes it the same either way.

This also updates esp-protocols (used for mdns) with a patch to
ensure that mangled names are the ones that collide.
(circuitpython.local collides but was causing cpy- to be mangled.)

Fixes #6869 again.